feat: remove npmrc and app-builder rename as _system
npmrc removed and app-builder rename as _system and moved to modules folder

Ref: #38
