--- .npmrc
@@ -1,3 +1,0 @@
-@wrappid:registry=https://npm.pkg.github.com/wrappid
-//npm.pkg.github.com/:_authToken=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
-registry=https://registry.npmjs.org
