--- package/.npmrc
@@ -1,3 +1,0 @@
-@wrappid:registry=https://npm.pkg.github.com/wrappid
-//npm.pkg.github.com/:_authToken=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
-registry=https://registry.npmjs.org
