+++ package/modules/_system/controllers.registry.ts
