+++ package/modules/_system/functions/businessEntity.get.helper.ts
@@ -1,5 +1,5 @@
-import { databaseProvider } from "../../index";
-import { databaseActions } from "../../index";
+import { databaseProvider } from "../../../index";
+import { databaseActions } from "../../../index";
 
 import {
   getEntitySchema,
