+++ package/modules/_system/functions/formSchema.helper.ts
@@ -1,6 +1,6 @@
-import { coreConstant } from "../../index";
-import { databaseProvider } from "../../index";
-import { databaseActions } from "../../index";
+import { coreConstant } from "../../../index";
+import { databaseProvider } from "../../../index";
+import { databaseActions } from "../../../index";
 
 // const { httpMethod, entityStatus } = coreConstant;
 import { getEntitySchema, getColumnsFromSchema } from "./businessEntity.helper";
