+++ package/modules/_system/utils/strings.utils.ts
