+++ package/registry/ControllersRegistry.ts
@@ -1,5 +1,5 @@
 import * as helperDatabase from "../database/helper.database";
-import AppBuilderControllersRegistry from "../app-builder/controllers.registry";
+import AppBuilderControllersRegistry from "../modules/_system/controllers.registry";
 const ControllersRegistry = {
   getDatabases: helperDatabase.getDatabases,
   getTables: helperDatabase.getTables,
